Implement undo/redo functionality
Undo/redo functionality is implemented using memento pattern. Each
memento snapshot records a delta of the editor itself, which is a vector
containing changes compared to previous version.

However, this version is only focused on Rust program API and does not
include any user interfaces, such as CLI or TUI.